Extract live chat event ownership
