+++ backend/app/codex_sdk_runner.py
@@ -2197,7 +2197,8 @@ async def run_codex_sdk_turn(
       # process group, so the transport dies mid-stream instead of
       # delivering turn/completed. Surfacing that as a provider failure is
       # both wrong and destructive: the raw string overwrites the stall note
-      # (chat._pause_note) published moments earlier, because error blocks
+      # (`chat_event_sink._pause_note`) published moments earlier, because
+      # error blocks
       # coalesce latest-wins and drop every events.ERROR_PASSTHROUGH_FIELDS
       # the new event omits — taking the note's one-tap Resume with it and
       # leaving the owner an unexplained error and no way back.
